% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AllClasses.R, R/DataGeNET_Psy-extract.R, R/DataGeNET_Psy-extractSentences.R, R/DataGeNET_Psy-ndisease.R, R/DataGeNET_Psy-ngene.R
\docType{class}
\name{DataGeNET.Psy-class}
\alias{DataGeNET.Psy-class}
\alias{DataGeNET.Psy-methods}
\alias{extract,DataGeNET.Psy-method}
\alias{extractSentences,DataGeNET.Psy-method}
\alias{ndisease,DataGeNET.Psy-method}
\alias{ngene,DataGeNET.Psy-method}
\title{Class DataGeNET.Psy}
\usage{
\S4method{extract}{DataGeNET.Psy}(object)

\S4method{extractSentences}{DataGeNET.Psy}(object, disorder, verbose = FALSE)

\S4method{ndisease}{DataGeNET.Psy}(object)

\S4method{ngene}{DataGeNET.Psy}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}}

\item{disorder}{A disorder to check if any evidence exists.}

\item{verbose}{If set to \code{TRUE} informative messages are show.}
}
\value{
A \code{data.frame} containing the raw result from PsyGeNET

The number of unique diseases

The number of unique genes
}
\description{
Class \code{DataGeNET.Psy} is the basic object use in \code{psygenet2r} 
package. It is the main data container to using the different functions to
query PsyGeNET database and generate teir output.
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: extract raw data

\item \code{extractSentences}: get sentences or evidences

\item \code{ndisease}: get number of diseases

\item \code{ngene}: get number of genes
}}
\section{Slots}{

\describe{
\item{\code{type}}{Character containing \code{'gene'} of \code{'disease'}. It is 
used to eprform the correct query to PsyGeNET.}

\item{\code{search}}{Character containing \code{'single'} of \code{'list'}.It is 
used to eprform the correct query to PsyGeNET.}

\item{\code{database}}{Character containing the name of the database that will be
queried. It can take the values \code{'MODELS'} to use Comparative 
Toxigenomics Database, data from mouse and rat; \code{'GAD'} to use Genetic 
Association Database; \code{'CTD'} to use Comparative Toxigenomics Database, 
data from human; \code{'PsyCUR'} to use Psychiatric disorders Gene 
association manually curated; \code{'CURATED'} to use Human, manually 
curated databases (PsyCUR and CTD); or \code{'ALL'} to use all these 
databases.}

\item{\code{term}}{Charcter with the term(s) to search into the database(s).}

\item{\code{qresult}}{\code{data.frame} with the obtained result}
}}
\examples{
extract(qr)[1:2, ] # get internat data.frame
ndisease(qr)
}
\seealso{
psygenetGene, psygenetDisease, DataGeNET.Psy-methods
}

