% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{ndisease}
\alias{extractSentences}
\alias{ndisease}
\alias{ngene}
\title{Getter from \code{DataGeNET.Psy}.}
\usage{
ndisease(object)

ngene(object)

extractSentences(object, disorder, verbose)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}.}

\item{disorder}{A disorder to check if any evidence exists.}

\item{verbose}{If set to \code{TRUE} informative messages are show.}
}
\value{
The number of unique genes

A data frame showing the evidence.
}
\description{
Obtain the number of unique diseases in a \code{DataGeNET.Psy}.

Obtain the number of unique genes in a \code{DataGeNET.Psy}.

Internally, PsyGeNET uses a series of collected evidences from public
literature. The internal table of a \code{DataGeNET.Psy} object 
can contains this information. The method \code{extractSentences} 
allows to extrat this information.
}
\examples{
ngene(qr)
extractSentences(qr, "Depression")
}

