% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedPD.R
\name{enrichedPD}
\alias{enrichedPD}
\title{Enrichment of a user's input (genes) in PsyGeNET's diseases.}
\usage{
enrichedPD(gene, database = "ALL", verbose = FALSE, warnings = FALSE)
}
\arguments{
\item{gene}{Name or vector of names (that can be both code or uml) to 
specific genes from PsyGeNET.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'psycur15'} to use data validated by experts for first release 
of PsyGeNET; \code{'psycur16'} to use data validated by experts for second 
release of PsyGeNET; or \code{'ALL'} to use both databases. 
Default \code{'ALL'}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to not see
the warnings.}
}
\value{
A \code{data.frame} with the enricment at each Psychiatric Disorder
}
\description{
Test the enrichment of a given gene list on Psychiatric Disorders from 
PsyGeNET.
}
\examples{
enrichedPD(c("ADCY2", "AKAP13", "ANK3"), "ALL")
}
