% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataGeNET_Psy-extract.R, R/JaccardIndexPsy-extract.R
\docType{methods}
\name{extract}
\alias{extract}
\title{Raw data from \code{DataGeNET.Psy} and \code{JaccardIndexPsy}.}
\usage{
\S4method{extract}{DataGeNET.Psy}(object)

\S4method{extract}{JaccardIndexPsy}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}}

\item{object}{Object of class \code{JaccardIndexPsy}}
}
\value{
A \code{data.frame} containing the raw result from PsyGeNET

A \code{data.frame} with the result Jaccard Index for each disease.
}
\description{
Obtain the raw data from a PsyGeNET's query  stored in a
\code{DataGeNET.Psy} object or the raw data with all the Jaccard Index for
the disease of interest of an \code{JaccardIndexPsy} object.

Obtain the raw data from a PsyGeNET's query  stored in a
\code{DataGeNET.Psy} object or the raw data with all the Jaccard Index for
the disease of interest of an \code{JaccardIndexPsy} object.
}
\examples{
if (require(psygenet2r)){
 extract(qr)[1:2, ] # get internat data.frame
}
\dontrun{
#Being x an JaccardIndexPsy
extract(x)
}
}

