% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R, R/DataGeNET_Psy-extract.R,
%   R/JaccardIndexPsy-extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract,DataGeNET.Psy-method}
\alias{extract,JaccardIndexPsy-method}
\title{Raw data from \code{DataGeNET.Psy} and \code{JaccardIndexPsy}.}
\usage{
extract(object, ...)

\S4method{extract}{DataGeNET.Psy}(object)

\S4method{extract}{JaccardIndexPsy}(object, order.cl = "pval", ...)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy} or \code{JaccardIndexPsy}}

\item{...}{NO USED}

\item{order.cl}{Order resulting \code{data.frame} by the name of this column.}
}
\value{
A \code{data.frame} containing the raw result from PsyGeNET or
a \code{data.frame} with the result Jaccard Index for each disease.

A \code{data.frame} containing the raw result from PsyGeNET

A \code{data.frame} with the result Jaccard Index for each disease.
}
\description{
Obtain the raw data from a PsyGeNET's query  stored in a 
\code{DataGeNET.Psy} object or the raw data with all the Jaccard Index for 
the disease of interest of an \code{JaccardIndexPsy} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataGeNET.Psy}: Extract function for DataGeNET.Psy

\item \code{JaccardIndexPsy}: Extract function for JacardIndexPsy
}}

\examples{
data(qr)
extract(qr)[1:2, ] # get internat data.frame
\dontrun{
#Being x an JaccardIndexPsy
extract(x)
}
}
