% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataGeNET_Psy-extractSentences.R
\docType{methods}
\name{extractSentences,DataGeNET.Psy-method}
\alias{extractSentences,DataGeNET.Psy-method}
\title{Method to obtain the evidences from a \code{DataGeNET.Psy} object.}
\usage{
\S4method{extractSentences}{DataGeNET.Psy}(input, disorder, verbose = FALSE)
}
\arguments{
\item{input}{A \code{DataGeNET.Psy} object.}

\item{disorder}{A disorder to check if any evidence exists.}

\item{vervose}{If set to \code{TRUE} informative messages are show.}
}
\value{
A data frame showing the evidence.
}
\description{
Internally, PsyGeNET uses a series of collected evidences from public
literature. The internal table of a \code{DataGeNET.Psy} object
can contains this information. The method \code{extractSentences}
allows to extrat this information.
}
\examples{
extractSentences(qr, "Depression")
}

