% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUMLs.R
\name{getUMLs}
\alias{getUMLs}
\title{Query PsyGeNET for given gene(s) and generates an \code{DataGeNET.Psy}}
\usage{
getUMLs(word, database = "ALL")
}
\arguments{
\item{word}{Disese to convert to UMLS using PsyGeNET database.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'psycur15'} to use data validated by experts for first release 
of PsyGeNET; \code{'psycur16'} to use data validated by experts for second 
release of PsyGeNET; or \code{'ALL'} to use both databases. 
Default \code{'ALL'}.}
}
\value{
The corresponding UMLs for the input disease/s
}
\description{
Given the name of one or multiple gene and retrives their information 
from PsyGeNET and creates an object of type \code{DataGeNET.Psy}.
}
\examples{
umls <- getUMLs( word = "depressive", database = "ALL" )
}

