% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardEstimation.R
\name{jaccardEstimation}
\alias{jaccardEstimation}
\title{Calculation of the Jaccard Index between ideseases}
\usage{
jaccardEstimation(pDisease, sDisease, database = "ALL", nboot = 100,
  ncores = 1, verbose = FALSE)
}
\arguments{
\item{pDisease}{vector of diseases, vector of genes, vector of main 
psychiatric disorder.}

\item{sDisease}{vector of diseases, vector of genes, vector of main 
psychiatric disorder. Only necessary when comparing genes vs. diseases,
genes vs. main psychiatric disorders or diseases vs. main psychiatric 
disorders. To compare multiple diseases only use \code{pDisease}.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'psycur15'} to use data validated by experts for first release 
of PsyGeNET; \code{'psycur16'} to use data validated by experts for second 
release of PsyGeNET; or \code{'ALL'} to use both databases.}

\item{nboot}{Number of iterations sued to compute the pvalue associted
to the calculated Jaccard Index (default 100).}

\item{ncores}{Number of cores used to calculate the pvalue associated to
the computed Jaccard Index (default 1).}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
An object of class \code{JaccardIndexPsy} with the computed
calculation of the JaccardIndex.
}
\description{
This function is able to calculate the Jacard Index between: 1. muliple
disases, 2. a set og genes and multiple diseases, 3. a set of genes and
multiple main psychiatric disorders and 4. multiple diseases and multiple
main psychiatric disorders.
}
\details{
Warning: The main psychiatric disorders are understood as a single set
of genes composed by the genes of all the diseases that the main 
psychiatric disorder cotains.
}
\examples{
ji <- jaccardEstimation( c( "COMT", "CLOCK", "DRD3" ), "umls:C0005586", "ALL" )
}
