% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AllClasses.R, R/JaccardIndexPsy-extract.R
\docType{class}
\name{JaccardIndexPsy-class}
\alias{JaccardIndexPsy-class}
\alias{JaccardIndexPsy-methods}
\alias{extract,JaccardIndexPsy-method}
\title{Class JaccardIndexPsy}
\usage{
\S4method{extract}{JaccardIndexPsy}(object)
}
\arguments{
\item{object}{Object of class \code{JaccardIndexPsy}}
}
\value{
A \code{data.frame} with the result Jaccard Index for each disease.
}
\description{
Class \code{JaccardIndexPsy} is theresult of the process to look for a 
Jaccard Index between muliple diseases in \code{psygenet2r} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: extract raw calulated values
}}
\section{Slots}{

\describe{
\item{\code{nit}}{Number of iterations to calculate the estimated Jaccard index}

\item{\code{type}}{Slot to save type of query (disease-disease, gene-disease)}

\item{\code{table}}{\code{data.frame} containing the result table of Jaccard indexes}

\item{\code{i1}}{[internal use] vector with names of first component}

\item{\code{i2}}{[internal use] vector with names of second component}
}}
\examples{
\dontrun{
#Being x an JaccardIndexPsy
extract(x)
}
}
\seealso{
psygenetGene, psygenetDisease, JaccardIndexPsy-methods
}

