% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeNET_Psy-ndisease.R
\docType{methods}
\name{ndisease}
\alias{ndisease}
\title{Getter from \code{DataGeNET.Psy}.}
\usage{
\S4method{ndisease}{DataGeNET.Psy}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}}
}
\value{
The number of unique diseases
}
\description{
Obtain the number of unique diseases in a \code{DataGeNET.Psy}.
}
\examples{
if (require(psygenet2r)){
 ndisease(qr)
}
}

