% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataGeNET_Psy-ngene.R
\docType{methods}
\name{ngene}
\alias{ngene}
\title{Getter from \code{DataGeNET.Psy}.}
\usage{
\S4method{ngene}{DataGeNET.Psy}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}}
}
\value{
The number of unique genes
}
\description{
Obtain the number of unique genes in a \code{DataGeNET.Psy}.
}
\examples{
if (require(psygenet2r)){
 ngene(qr)
}
}

