% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R, R/DataGeNET_Psy-ngene.R
\name{ngene}
\alias{ngene}
\alias{ngene,DataGeNET.Psy-method}
\title{Getter from \code{DataGeNET.Psy}.}
\usage{
ngene(object)

\S4method{ngene}{DataGeNET.Psy}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}.}
}
\value{
The number of unique genes

The number of unique genes
}
\description{
Obtain the number of unique genes in a \code{DataGeNET.Psy}.
}
\section{Functions}{
\itemize{
\item \code{ngene,DataGeNET.Psy-method}: Get number of genes
}}

\examples{
data(qr)
ngene(qr)
}
