% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pantherGraphicCorrect.R
\name{pantherGraphic}
\alias{pantherGraphic}
\title{Query PsyGeNET for given genes and creates a representation in base of
their panther-class}
\usage{
pantherGraphic(x, database = "CURATED", score, verbose = FALSE)
}
\arguments{
\item{x}{Vector of genes of interest of \code{DataGeNET.Psy} resulting of
\code{psyegnetDisease}.}

\item{database}{Name of the database that will be queried. It can take the
values \code{'MODELS'} to use Comparative Toxigenomics Database, data from
mouse and rat; \code{'GAD'} to use Genetic Association Database; \code{'CTD'}
to use Comparative Toxigenomics Database, data from human; \code{'PsyCUR'} to
use Psychiatric disorders Gene association manually curated; \code{'CURATED'}
to use Human, manually curated databases (PsyCUR and CTD); or \code{'ALL'}
to use all these databases. Default \code{'CURATED'}.}

\item{score}{threshold to take into account a gene in the analysis}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
A plot for a \code{DataGeNET.Psy} in terms of the panther-class.
}
\description{
Given a vector of genes of interest (or using a DataGeNET.Psy object), this
function creates a representation of a the panther-class these genes belongs
to.
}
\examples{
d.alch <- pantherGraphic( c( "COMT", "CLOCK", "DRD3" ), "CURATED" )
}

