% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataGeNET_Psy-plot.R, R/JaccardIndexPsy-plot.R
\docType{methods}
\name{plot}
\alias{plot}
\title{Plots the content of a \code{DataGeNET.Psy} or \code{JaccardIndexPsy}
object.}
\usage{
\S4method{plot}{DataGeNET.Psy,ANY}(x, y, ...)

\S4method{plot}{JaccardIndexPsy,ANY}(x, cutOff, zero.remove = TRUE,
  noTitle = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{DataGeNET.Psy}}

\item{...}{NOT USED}

\item{cutOff}{Number to filter the shown results.}

\item{zero.remove}{By deffault \code{TRUE}. It removes those relations
with a Jaccard Index of 0.}

\item{noTitle}{By default \code{FALSE}. If set to true no title will be
added to the plot.}

\item{verbose}{By default \code{FALSE}. If set to \code{TRUE} information
on the drawing process will be shown.}

\item{layout}{Function to design the location of the different nodes. By
default \code{layout.fruchterman.reingold} from \code{igraph} is used.}

\item{type}{Type of the drawn chart. By default it is \code{"disease"} but
it also can be \code{"individual"}, \code{"disease"},
\code{"disease class"}, \code{"venn"}, \code{"barplot"},
\code{"heatmapGenes"}, \code{"heatmapScore"} or \code{"heatmap"}.}

\item{x}{Object of class \code{JaccardIndexPsy}.}

\item{verbose}{By default \code{FALSE}. If set to \code{TRUE} information
on the drawing process will be shown.}
}
\value{
A plot for \code{DataGeNET.Psy}.

A plot for \code{JaccardIndexPsy}.
}
\description{
This functions llows to create a variety of plots for \code{DataGeNEt.Psy}
and \code{JaccardIndexPsy} objects.

This functions llows to create a variety of plots for \code{DataGeNEt.Psy}
and \code{JaccardIndexPsy} objects.
}
\examples{
if (require(psygenet2r)){
 plot(x) # for all-disease plot
 plot(qr, type = 'disease class') # for MPI plot
}
\dontrun{
#Being x an JaccardIndexPsy
qr <- plot(x)
}
}

