% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygenetDisease.R
\name{psygenetDisease}
\alias{psygenetDisease}
\title{Query PsyGeNET for given disease(s) and generates an \code{DataGeNET.Psy}}
\usage{
psygenetDisease(disease, database = "CURATED", score = c(">", 0),
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{disease}{Name or vector of names (that can be both code or uml) to 
specific diseases from PsyGeNET. The diseases non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'MODELS'} to use Comparative Toxigenomics Database, data from 
mouse and rat; \code{'GAD'} to use Genetic Association Database; \code{'CTD'}
to use Comparative Toxigenomics Database, data from human; \code{'PsyCUR'} to
use Psychiatric disorders Gene association manually curated; \code{'CURATED'}
to use Human, manually curated databases (PsyCUR and CTD); or \code{'ALL'} 
to use all these databases. Default \code{'CURATED'}.}

\item{score}{A vector with two elements: 1) character with greather 
\code{'>'} or with lower \code{'<'} meaing greather or equal and lower or
equal; 2) the score to be compared. By default: \code{c('>', 0)}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given the name of one or multiple diseases and retrives their information 
from PsyGeNET and creates an object of type \code{DataGeNET.Psy}.
}
\examples{
d.alch <- psygenetDisease( "Alcoholism", "CURATED" )
}

