% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygenetDiseaseSentences.R
\name{psygenetDiseaseSentences}
\alias{psygenetDiseaseSentences}
\title{Query PsyGeNET for given disease(s) and extract the pmids sentences that report 
a gene-disease asssociation.}
\usage{
psygenetDiseaseSentences(diseaseList, database, verbose = FALSE)
}
\arguments{
\item{diseaseList}{Name or vector of names (that can be both code or uml) to 
specific diseases from PsyGeNET. The diseases non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'MODELS'} to use Comparative Toxigenomics Database, data from 
mouse and rat; \code{'GAD'} to use Genetic Association Database; \code{'CTD'}
to use Comparative Toxigenomics Database, data from human; \code{'PsyCUR'} to
use Psychiatric disorders Gene association manually curated; \code{'CURATED'}
to use Human, manually curated databases (PsyCUR and CTD); or \code{'ALL'} 
to use all these databases. Default \code{'CURATED'}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given a disease or a disease list, retrives the pmids and sentences for each 
gene-disease association from PsyGeNET and creates an object of type 
\code{DataGeNET.Psy}.
}
\examples{
diseasesOfInterest <- c( "Bipolar Disorder","Depressive Disorder, Major" )
psyDisSen <- psygenetDiseaseSentences( diseaseList = diseasesOfInterest,
                                       database = "ALL" )
}

