% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygenetGene.R
\name{psygenetGene}
\alias{psygenetGene}
\title{Query PsyGeNET for given gene(s) and generates an \code{DataGeNET.Psy}}
\usage{
psygenetGene(gene, database = "ALL", score = c(">", 0), check = FALSE,
  hostMart = "www.ensembl.org", biomart = "ENSEMBL_MART_ENSEMBL",
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{gene}{Name or vector of names (that can be both code or symbol) to 
specific genes from PsyGeNET. The genes non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'psycur15'} to use data validated by experts for first release 
of PsyGeNET; \code{'psycur16'} to use data validated by experts for second 
release of PsyGeNET; or \code{'ALL'} to use both databases. 
Default \code{'ALL'}.}

\item{score}{A vector with two elements: 1) character with greather 
\code{'>'} or with lower \code{'<'} meaing greather or equal and lower or}

\item{check}{By default \code{FALSE}. Change it to \code{TRUE} to 
validate the genes to biomart.}

\item{hostMart}{The URL of Biomart to be used.}

\item{biomart}{By default \code{'ENSEMBL_MART_ENSEMBL'}. The mart of biomart used to 
check genes}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to not see
the warnings.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given the name of one or multiple gene and retrives their information 
from PsyGeNET and creates an object of type \code{DataGeNET.Psy}.
}
\examples{
d.alch <- psygenetGene( "ALDH2", "ALL", check = FALSE )
}

