% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/psygenetGeneSentences.R
\name{psygenetGeneSentences}
\alias{psygenetGeneSentences}
\title{Query PsyGeNET for given gene(s) and extract the pmids sentences that report
a gene-disease asssociation.}
\usage{
psygenetGeneSentences(geneList, database = "CURATED", verbose = FALSE)
}
\arguments{
\item{geneList}{Name or vector of names (that can be both code or symbol) to
specific genes from PsyGeNET. The genes non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the
values \code{'MODELS'} to use Comparative Toxigenomics Database, data from
mouse and rat; \code{'GAD'} to use Genetic Association Database; \code{'CTD'}
to use Comparative Toxigenomics Database, data from human; \code{'PsyCUR'} to
use Psychiatric disorders Gene association manually curated; \code{'CURATED'}
to use Human, manually curated databases (PsyCUR and CTD); or \code{'ALL'}
to use all these databases. Default \code{'CURATED'}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given a gene or a gene list, retrives the pmids and sentences for each
gene-disease association from PsyGeNET and creates an object of type
\code{DataGeNET.Psy}.
}
\examples{
genesOfInterest <- c("PECR", "ADH1C", "CAST", "ERAP1", "PPP2R2B",
                     "ESR1", "GATA4", "CDH13")
psyGeneSen <- psygenetGeneSentences( geneList = genesOfInterest,
                                     database = "ALL")
}

