% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qr.R
\name{qr}
\alias{qr}
\title{\code{DataGeNET.Psy} obtained from quering PsyGeNET for gene '4852'.}
\format{The format is:
Formal class 'DataGeNET.Psy' [package "psygenet2r"] with 5 slots
  .. type    : chr "gene"
  .. search  : chr ""
  .. database: chr "ALL"
  .. term    : chr "4852"
  .. qresult :'data.frame'}
\source{
\url{http://psygenet.org}
}
\usage{
data("qr")
}
\value{
A \code{DataGeNET.Psy} object.
}
\description{
A dataset obtained from PsyGeNET after being queried with
\code{psygenetGene} usig the term '4852' on \code{"ALL"} databse.
}
\examples{
ngene(qr)
ndisease(qr)
}

