% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\docType{data}
\name{universe}
\alias{universe}
\title{Vector with gene universe for Jaccard Index}
\format{
An object of class \code{character} of length 8947.
}
\source{
\url{http://www.disgenet.org}
}
\usage{
universe
}
\value{
A \code{character} vector.
}
\description{
Vector with all the gene names from DisGeNET database
(http://www.disgenet.org) used as gene universe for
Jaccard Index computation.
}
\details{
data("universe", package = "psygenet2r")
}
\examples{
length(universe)
universe[1:10]
}
\keyword{datasets}
