% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R, R/DataGeNET_Psy-extractSentences.R
\docType{methods}
\name{extractSentences}
\alias{extractSentences}
\alias{extractSentences,DataGeNET.Psy-method}
\title{Method to obtain the evidences from a \code{DataGeNET.Psy} object.}
\usage{
extractSentences(object, disorder, verbose)

\S4method{extractSentences}{DataGeNET.Psy}(object, disorder, verbose = FALSE)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}.}

\item{disorder}{A disorder to check if any evidence exists.}

\item{verbose}{If set to \code{TRUE} informative messages are show.}
}
\value{
A data frame showing the evidence.
}
\description{
Internally, PsyGeNET uses a series of collected evidences from public
literature. The internal table of a \code{DataGeNET.Psy} object
can contains this information. The method \code{extractSentences}
allows to extrat this information.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataGeNET.Psy}: Get sentences or evidences
}}
\examples{
data(qr)
extractSentences(qr, "Depression")
}

