% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R,
%   R/DataGeNET_Psy-extractSentences.R
\docType{methods}
\name{extractSentences}
\alias{extractSentences}
\alias{extractSentences,DataGeNET.Psy-method}
\title{Method to obtain the sentences that support a gene-disease association 
from a \code{DataGeNET.Psy} object.}
\usage{
extractSentences(object, disorder, verbose)

\S4method{extractSentences}{DataGeNET.Psy}(object, disorder, verbose = FALSE)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}.}

\item{disorder}{The disorder of interest. Only those sentences supporting 
a gene-disease association with this specific disorder will be extracted.
Disorder must be provided by using "Disease Id" or "Disease Name". The 
"Disease Id" can be provided with or without the "uml" tag. Example of a
"Disease Name": "schizophrenia"; Example of a "Disease Id": "umls:C0036341",
that is equivalent of using "C0036341".}

\item{verbose}{If set to \code{TRUE} informative messages are show.}
}
\value{
A data frame showing the sentences.
}
\description{
PsyGeNET contains a list of sentences that support a gene-disease 
association from public literature. The internal table of a 
\code{DataGeNET.Psy} object contains this information. The method 
\code{extractSentences}  allows to extract those sentences that support a 
gene-disease association given a \code{DataGeNET.Psy} object and a disorder 
of interest.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataGeNET.Psy}: Get sentences or evidences
}}

\examples{
data(qr)
extractSentences(qr, "Depression")
}
