% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAttrPlot.R
\name{geneAttrPlot}
\alias{geneAttrPlot}
\title{Ploting the relation between genes and disease-categories}
\usage{
geneAttrPlot(x, type = "pie", verbose = FALSE)
}
\arguments{
\item{x}{Vector of genes of interest of \code{DataGeNET.Psy} resulting of
\code{psyegnetDisease}.}

\item{type}{Type of the drawn chart. By default it is \code{"pie"}. It can 
takes \code{"pie"} to plot a pie chart with the number of genes for each 
psychiatric category, \code{"disease category"} for visualizing a barplot 
with the total and specific number of genes for each psychiatric disorder, 
\code{"evidence index"} for a barplot showing for each psychiatric disorder
the number of gene-disease associations according to the Evidence index
and \code{"gene"} for visualizing a barplot with the total and specific 
number of diseases associated to each gene.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
A plot for a \code{DataGeNET.Psy} in terms of the panther-class.
}
\description{
Given a set of genes or a result of \code{psygenetGene} creates four types
of plots showing the relation of the genes with the disease's category
in psyGeNET.
}
\note{
The "Evidence Index" is gotten from PsyGeNET. For more information
about it and its calculation, pease visit \code{psygenet.org}.
}
