% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUMLs.R
\name{getUMLs}
\alias{getUMLs}
\title{Query PsyGeNET for given gene(s) and generates an \code{DataGeNET.Psy}}
\usage{
getUMLs(word, database = "CURATED")
}
\arguments{
\item{word}{Disese to convert to UMLS using PsyGeNET database.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'MODELS'} to use Comparative Toxigenomics Database, data from 
mouse and rat; \code{'GAD'} to use Genetic Association Database; \code{'CTD'}
to use Comparative Toxigenomics Database, data from human; \code{'PsyCUR'} to
use Psychiatric disorders Gene association manually curated; \code{'CURATED'}
to use Human, manually curated databases (PsyCUR and CTD); or \code{'ALL'} 
to use all these databases. Default \code{'CURATED'}.}
}
\value{
The corresponding UMLs for the input disease/s
}
\description{
Given the name of one or multiple gene and retrives their information 
from PsyGeNET and creates an object of type \code{DataGeNET.Psy}.
}
\examples{
umls <- getUMLs( "Alcoholism", "CURATED" )
}

