% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_AllClasses.R
\docType{class}
\name{JaccardIndexPsy-class}
\alias{JaccardIndexPsy-class}
\title{Class JaccardIndexPsy}
\description{
Class \code{JaccardIndexPsy} is theresult of the process to look for a
Jaccard Index between muliple diseases in \code{psygenet2r} package.
}
\section{Slots}{

\describe{
\item{\code{index}}{}

\item{\code{nit}}{Number of iterations to calculate the estimated Jaccard index}

\item{\code{type}}{}

\item{\code{table}}{\code{data.frame} containing the result table of Jaccard indexes}
}}
\seealso{
psygenetGene, psygenetDisease, JaccardIndexPsy-methods
}

