% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AllClasses.R
\docType{class}
\name{JaccardIndexPsy-class}
\alias{JaccardIndexPsy-class}
\title{Class JaccardIndexPsy}
\description{
Class \code{JaccardIndexPsy} is theresult of the process to look for a 
Jaccard Index between muliple diseases in \code{psygenet2r} package.
}
\section{Slots}{

\describe{
\item{\code{nit}}{Number of iterations to calculate the estimated Jaccard index}

\item{\code{type}}{Slot to save type of query (disease-disease, gene-disease)}

\item{\code{table}}{\code{data.frame} containing the result table of Jaccard indexes}

\item{\code{i1}}{[internal use] vector with names of first component}

\item{\code{i2}}{[internal use] vector with names of second component}
}}
\examples{
ji <- jaccardEstimation( c( "COMT", "CLOCK", "DRD3" ), "umls:C0005586", "ALL" )
}
\seealso{
psygenetGene, psygenetDisease, JaccardIndexPsy-methods
}

