% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_AllGenerics.R, R/DataGeNET_Psy-ndisease.R
\docType{methods}
\name{ndisease}
\alias{ndisease}
\alias{ndisease,DataGeNET.Psy-method}
\title{Getter from \code{DataGeNET.Psy}.}
\usage{
ndisease(object)

\S4method{ndisease}{DataGeNET.Psy}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.Psy}.}
}
\value{
The number of unique diseases
}
\description{
Obtain the number of unique diseases in a \code{DataGeNET.Psy}.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataGeNET.Psy}: Get number of diseases
}}
\examples{
data(qr)
ndisease(qr)
}

