% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeNET_Psy-plot.R
\docType{methods}
\name{plot,DataGeNET.Psy,ANY-method}
\alias{plot,DataGeNET.Psy,ANY-method}
\title{Plots the content of a \code{DataGeNET.Psy} or \code{JaccardIndexPsy} 
object.}
\usage{
\S4method{plot}{DataGeNET.Psy,ANY}(x, y,
  layout = igraph::layout.fruchterman.reingold, type = "GDA network",
  verbose = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{DataGeNET.Psy}}

\item{y}{NOT USED}

\item{layout}{Function to design the location of the different nodes. By 
default \code{layout.fruchterman.reingold} from \code{igraph} is used.}

\item{type}{Type of the drawn chart. By default it is \code{"GDA network"} 
but it also can be \code{"GDCA network"}, \code{"GDCA heatmap"}, 
\code{"GDA heatmap"} and \code{"publications"}. The first two are network
representations of the second two. While the last one draws a barplot
with the number of PMIDs between genes and diseases.}

\item{verbose}{By default \code{FALSE}. If set to \code{TRUE} information
on the drawing process will be shown.}

\item{...}{Passed to inner functions for different plots.}
}
\value{
A plot for \code{DataGeNET.Psy}.
}
\description{
This functions llows to create a variety of plots for \code{DataGeNEt.Psy}
and \code{JaccardIndexPsy} objects.
}
\note{
The "Evidence Index" is gotten from PsyGeNET. For more information
about it and its calculation, pease visit \code{psygenet.org}.
}
\examples{
data(qr)
plot(qr) # for all-disease plot
plot(qr, type = 'GDCA network') # for MPI plot
}
