% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeNET_Psy-plot.R
\docType{methods}
\name{plot,DataGeNET.Psy,ANY-method}
\alias{plot,DataGeNET.Psy,ANY-method}
\title{Plots the content of a \code{DataGeNET.Psy} or \code{JaccardIndexPsy} 
object.}
\usage{
\S4method{plot}{DataGeNET.Psy,ANY}(x, y,
  layout = igraph::layout.fruchterman.reingold, type = "disease",
  verbose = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{DataGeNET.Psy}}

\item{y}{NOT USED}

\item{layout}{Function to design the location of the different nodes. By 
default \code{layout.fruchterman.reingold} from \code{igraph} is used.}

\item{type}{Type of the drawn chart. By default it is \code{"disease"} but 
it also can be \code{"individual disease"}, \code{"disease"}, 
\code{"disease class"}, \code{"barplot"}, 
\code{"heatmapGenes"} or \code{"heatmap"}.}

\item{verbose}{By default \code{FALSE}. If set to \code{TRUE} information
on the drawing process will be shown.}

\item{...}{Passed to inner functions for different plots.}
}
\value{
A plot for \code{DataGeNET.Psy}.
}
\description{
This functions llows to create a variety of plots for \code{DataGeNEt.Psy}
and \code{JaccardIndexPsy} objects.
}
\examples{
data(qr)
plot(qr) # for all-disease plot
plot(qr, type = 'disease class') # for MPI plot
}

