% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygenetDisease.R
\name{psygenetDisease}
\alias{psygenetDisease}
\title{Query PsyGeNET for given disease(s) and generates an \code{DataGeNET.Psy}}
\usage{
psygenetDisease(disease, database = "ALL", score = c(">", 0),
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{disease}{Name or vector of names (that can be both code or uml) to 
specific diseases from PsyGeNET. The diseases non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'psycur15'} to use data validated by experts for first release 
of PsyGeNET; \code{'psycur16'} to use data validated by experts for second 
release of PsyGeNET; or \code{'ALL'} to use both databases. 
Default \code{'ALL'}.}

\item{score}{A vector with two elements: 1) character with greather 
\code{'>'} or with lower \code{'<'} meaing greather or equal and lower or
equal; 2) the score to be compared. By default: \code{c('>', 0)}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given the name of one or multiple diseases and retrives their information 
from PsyGeNET and creates an object of type \code{DataGeNET.Psy}.
}
\examples{
d.sch <- psygenetDisease( "schizophrenia", "ALL" )
}

