% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygenetDiseaseSentences.R
\name{psygenetDiseaseSentences}
\alias{psygenetDiseaseSentences}
\title{Query PsyGeNET for given disease(s) and extract the pmids sentences that report 
a gene-disease asssociation.}
\usage{
psygenetDiseaseSentences(diseaseList, database = "ALL", verbose = FALSE)
}
\arguments{
\item{diseaseList}{Name or vector of names (that can be both code or uml) to 
specific diseases from PsyGeNET. The diseases non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'psycur15'} to use data validated by experts for first release 
of PsyGeNET; \code{'psycur16'} to use data validated by experts for second 
release of PsyGeNET; or \code{'ALL'} to use both databases. 
Default \code{'ALL'}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given a disease or a disease list, retrives the pmids and sentences for each 
gene-disease association from PsyGeNET and creates an object of type 
\code{DataGeNET.Psy}.
}
\examples{
diseasesOfInterest <- c( "Bipolar Disorder","Depressive Disorder, Major" )
psyDisSen <- psygenetDiseaseSentences( diseaseList = diseasesOfInterest,
                                       database = "ALL" )
}
