% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygenetGene.R
\name{psygenetGene}
\alias{psygenetGene}
\title{Query PsyGeNET for given gene(s) and generates an \code{DataGeNET.Psy}}
\usage{
psygenetGene(gene, database = "CURATED", score = c(">", 0), check = TRUE,
  hostMart = "www.ensembl.org", verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{gene}{Name or vector of names (that can be both code or symbol) to 
specific genes from PsyGeNET. The genes non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'MODELS'} to use Comparative Toxigenomics Database, data from 
mouse and rat; \code{'GAD'} to use Genetic Association Database; \code{'CTD'}
to use Comparative Toxigenomics Database, data from human; \code{'PsyCUR'} to
use Psychiatric disorders Gene association manually curated; \code{'CURATED'}
to use Human, manually curated databases (PsyCUR and CTD); or \code{'ALL'} 
to use all these databases. Default \code{'CURATED'}.}

\item{score}{A vector with two elements: 1) character with greather 
\code{'>'} or with lower \code{'<'} meaing greather or equal and lower or}

\item{check}{By default \code{TRUE}. Change it to \code{FALSE} to not
validate the genes.}

\item{hostMart}{The URL of Biomart to be used.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to not see
the warnings.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given the name of one or multiple gene and retrives their information 
from PsyGeNET and creates an object of type \code{DataGeNET.Psy}.
}
\examples{
d.alch <- psygenetGene( "ALDH2", "CURATED", check = FALSE )
}

