% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygenetGeneSentences.R
\name{psygenetGeneSentences}
\alias{psygenetGeneSentences}
\title{Query PsyGeNET for given gene(s) and extract the pmids sentences that report 
a gene-disease asssociation.}
\usage{
psygenetGeneSentences(geneList, database = "ALL", verbose = FALSE)
}
\arguments{
\item{geneList}{Name or vector of names (that can be both code or symbol) to 
specific genes from PsyGeNET. The genes non existing in PsyGeNET will
be removed from the output.}

\item{database}{Name of the database that will be queried. It can take the 
values \code{'psycur15'} to use data validated by experts for first release 
of PsyGeNET; \code{'psycur16'} to use data validated by experts for second 
release of PsyGeNET; or \code{'ALL'} to use both databases. 
Default \code{'ALL'}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}
}
\value{
An object of class \code{DataGeNET.Psy}
}
\description{
Given a gene or a gene list, retrives the pmids and sentences for each 
gene-disease association from PsyGeNET and creates an object of type 
\code{DataGeNET.Psy}.
}
\examples{
genesOfInterest <- c("PECR", "ADH1C", "CAST", "ERAP1", "PPP2R2B",
                     "ESR1", "GATA4", "CDH13")
psyGeneSen <- psygenetGeneSentences( geneList = genesOfInterest,
                                     database = "ALL")
}

