% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topAnatEnrichment.R
\name{topAnatEnrichment}
\alias{topAnatEnrichment}
\title{Enrichment of a user's input (genes) in anatomical terms (TopAnat).}
\usage{
topAnatEnrichment(gene, datatype = c("rna_seq", "affymetrix", "est",
  "in_situ"), statistic = "fisher", cutOff = 1, verbose = FALSE,
  warnings = FALSE)
}
\arguments{
\item{gene}{Name or vector of names (that can be both code or uml) to 
specific genes from PsyGeNET.}

\item{datatype}{It can take the values \code{'rna_seq'}, \code{'affymetrix'}, 
\code{"est"} or \code{"in situ"}. 
Default \code{c("rna_seq","affymetrix","est","in_situ")}.}

\item{statistic}{By default it is \code{"fisher"}. But it can be changed to
\code{"ks"}, \code{"t"}, \code{"globaltest"}, \code{"sum"} or 
\code{"ks.ties"}. All from \code{\link[topGO]{runTest}}.}

\item{cutOff}{Default \code{1}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to not see
the warnings.}
}
\value{
A \code{data.frame} with the enrichment results
}
\description{
Test the enrichment of a given gene list on Psychiatric Disorders from 
PsyGeNET.
}
\examples{
\dontrun{
topAnatEnrichment(gene=c("ADCY2", "AKAP13", "ANK3"))
}
}
